smartwaste/requirements.txt
flask==2.3.3
gunicorn==21.2.0
ultralytics==8.0.196
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.0.1
werkzeug==2.3.7
python-dotenv==1.0.0
pymongo==4.6.0
dnspython==2.4.2